feat[judge-handler]: add submission-related methods

+++ src/main/java/com/example/codeE/model/exercise/CodeExercise.java
@@ -39,7 +39,7 @@ public class CodeExercise extends Exercise {
 
     @Field
     @NotNull(message = "Memory limit is required")
-    private Integer memoryLimit;
+    private Double memoryLimit;
 
     @Field("short_circuit")
     private Boolean shortCircuit;
